Windows support
Windows complains about non-exported base types across a DLL boundary.

Is this a good error? Is the linux link model capable of handling this, or would this have accidentally ruined RTTI beforehand? Unsure. Does seem a bit weird to need to export something that is essentially a header file ... but then I don't know how the platforms manage the vtable.